slider seek

+++ app/src/main/java/cn/x/service/PlayerController.kt
@@ -39,7 +39,7 @@ interface PlayerController {
     fun prev()
 
     @MainThread
-    fun seekTo(msec: Int)
+    fun seekTo(msec: Long)
 
     @MainThread
     fun getAudioSessionId(): Int
